move client updates to ForeignClient
